third_party/blink/web_tests/virtual/coop-restrict-properties/external/wpt/html/cross-origin-opener-policy/tentative/restrict-properties/iframe-popup-about-blank.https.window-expected.txt

This is a testharness.js-based test.
FAIL iframe-popup-about-blank assert_equals: Is popup handle null? expected "false" but got "true"
Harness: the test ran to completion.
